requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
attrs==18.2.0
cffi==1.11.5
chardet==3.0.4
discord.py==0.16.12
idna==2.8
multidict==4.5.2
pycparser==2.19
six==1.12.0
websockets==3.4
yarl==1.3.0
